requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
google-generativeai==0.4.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
webdriver-manager==4.0.1